% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_document.R
\docType{data}
\name{Goosefish}
\alias{Goosefish}
\title{Goosefish: Northern Management Region (for ML)}
\format{An object of class \code{\linkS4class{MLZ_data}}.}
\usage{
Goosefish
}
\description{
Mean lengths and life history for goosefish.
}
\examples{
data(Goosefish); Goosefish

}
\references{
Gedamke, T. and Hoenig, J.M. 2006. Estimating mortality from mean length data in
nonequilibrium situations, with application to the assessment of goosefish.
Transactions of the American Fisheries Society 135:476-487.
}
\keyword{datasets}
