% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_document.R
\docType{data}
\name{MuttonSnapper}
\alias{MuttonSnapper}
\title{Puerto Rico Mutton Snapper}
\format{An object of class \code{\linkS4class{MLZ_data}}.}
\usage{
MuttonSnapper
}
\description{
Mean lengths, CPUE, and life history for Puerto Rico mutton snapper.
}
\examples{
data(MuttonSnapper)

}
\references{
Huynh, Q.C., Gedamke, T., Porch, C.E., Hoenig, J.M., Walter, J.F, Bryan, M, and Brodziak, J.
In revision. Estimating Total Mortality Rates from Mean Lengths and Catch Rates in Non-equilibrium
Situations. Transactions of the American Fisheries Society.
}
\keyword{datasets}
