% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_document.R
\docType{data}
\name{Nephrops}
\alias{Nephrops}
\title{Male Nephrops FU 28-29 (for MLeffort)}
\format{An object of class \code{\linkS4class{MLZ_data}}.}
\usage{
Nephrops
}
\description{
An S4 object containing length and effort time series and life history parameters of male Nephrops in FU 28-29.
}
\examples{
data(Nephrops)

}
\references{
Then, A.Y, Hoenig, J.M, and Huynh, Q.C. In revision. Estimating fishing and natural mortality rates,
and catchability coefficient, from a series of observations on mean length and fishing effort.
ICES Journal of Marine Science.
}
\keyword{datasets}
