% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{plot,MLZ_data,ANY-method}
\alias{plot,MLZ_data,ANY-method}
\title{\code{plot} method for S4 class \code{MLZ_data}}
\usage{
\S4method{plot}{MLZ_data,ANY}(x, ggplot_layer = NULL)
}
\arguments{
\item{x}{An object of class \code{MLZ_data}.}

\item{ggplot_layer}{Layers to add to ggplot2 plot.}
}
\description{
Plots annual length frequencies from slot \code{Len_matrix} or \code{Len_df}. If there are data in 
both slots, \code{Len_matrix} is preferentially plotted.
}
\examples{
data(Nephrops)
plot(Nephrops)

}
