% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{plot,MLZ_model-method}
\alias{plot,MLZ_model-method}
\title{\code{plot} method for S4 class \code{MLZ_model}}
\usage{
\S4method{plot}{MLZ_model}(x, residuals = TRUE)
}
\arguments{
\item{x}{An object of class \code{MLZ_model}.}

\item{residuals}{logical; whether a plot of residuals will also be produced.}
}
\description{
Plots time series of observed and predicted data from an object of class \code{MLZ_model}.
}
\examples{
data(Goosefish)
goose.model <- ML(Goosefish, ncp = 2, grid.search = FALSE, figure = FALSE)
plot(goose.model)
}
