% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_ML}
\alias{profile_ML}
\title{Grid search for the mean length estimator}
\usage{
profile_ML(MLZ_data, ncp, stZ = 0.5, min.time = 3, parallel = FALSE,
  figure = TRUE, color = TRUE)
}
\arguments{
\item{MLZ_data}{An object of class \code{MLZ_data}.}

\item{ncp}{The number of change points.}

\item{stZ}{The starting value of total mortality rate used in the grid search.}

\item{min.time}{The minimum number of years between change points. Only used if \code{ncp > 1}.}

\item{parallel}{Whether grid search is performed using parallel processing.}

\item{figure}{If \code{TRUE}, creates a plot of the likelihood over the grid search. Only used
if \code{ncp = 1} or \code{2}.}

\item{color}{If \code{TRUE}, creates a color plot for the likelihood surface. Only used if
\code{ncp = 2}.}
}
\description{
A grid search is performed over the time series, which can be used to identify local and global minima. A
plot of the likelihood surface is also created similar to Figure 6 of Gedamke and Hoenig (2006).
}
\examples{
data(Goosefish)
profile_ML(Goosefish, ncp = 1)
profile_ML(Goosefish, ncp = 2)

}
