% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_MLCR}
\alias{profile_MLCR}
\title{Grid search for the mean length with catch rate estimator}
\usage{
profile_MLCR(MLZ_data, ncp, CPUE.type = c("NPUE", "WPUE"),
  loglikeCPUE = c("lognormal", "normal"), spawn = c("continuous", "annual"),
  stZ = 0.5, min.time = 3, parallel = ifelse(ncp > 2, TRUE, FALSE),
  figure = TRUE, color = TRUE)
}
\arguments{
\item{MLZ_data}{An object of class \code{MLZ_data}.}

\item{ncp}{The number of change points.}

\item{CPUE.type}{Indicates whether CPUE time series is abundance or biomass based.}

\item{loglikeCPUE}{Indicates whether the log-likelihood for the CPUE will be lognormally or
normally distributed.}

\item{spawn}{Whether 'continuous' or 'annual' (pulse) spawning is modeled.}

\item{stZ}{The starting value of total mortality rate used in the grid search.}

\item{min.time}{The minimum number of years between change points. Only used if \code{ncp > 1}.}

\item{parallel}{Whether the grid search is performed with parallel processing.}

\item{figure}{If \code{TRUE}, creates a plot of the likelihood over the grid search. Only used
if \code{ncp = 1} or \code{2}.}

\item{color}{If \code{TRUE}, creates a color plot for the likelihood surface. Only used if
\code{ncp = 2}.}
}
\description{
A grid search is performed over the time series. Can be used to identify local and global minima.
}
\examples{
data(MuttonSnapper)
profile_MLCR(MuttonSnapper, ncp = 1)
}
