% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_document.R
\docType{data}
\name{PRSnapper}
\alias{PRSnapper}
\title{Puerto Rico Snapper}
\format{A list containing objects of class \code{\linkS4class{MLZ_data}}.}
\usage{
PRSnapper
}
\description{
Mean lengths and life history for 3 species in the Puerto Rico Deepwater Snapper Complex (Unit 1):
silk snapper, blackfin snapper, and vermilion snapper.
}
\examples{
data(PRSnapper)

}
\references{
Huynh, Q.C, Gedamke, T., Hoenig, J.M, and Porch C. In press. Multispecies Extensions to a Nonequilibrium
Length-Based Mortality Estimator. Marine and Coastal Fisheries.
}
\keyword{datasets}
