% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_document.R
\docType{data}
\name{SilkSnapper}
\alias{SilkSnapper}
\title{Silk Snapper}
\format{A data frame.}
\usage{
SilkSnapper
}
\description{
Length observed from the Puerto Rico Silk Snapper handline fishery.
}
\examples{
data(SilkSnapper)
new("MLZ_data", Len_df = SilkSnapper)

}
\references{
Huynh, Q.C, Gedamke, T., Hoenig, J.M, and Porch C. In press. Multispecies Extensions to a Nonequilibrium
Length-Based Mortality Estimator. Marine and Coastal Fisheries.
}
\keyword{datasets}
