% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{summary,MLZ_data-method}
\alias{summary,MLZ_data-method}
\title{\code{summary} method for S4 class \code{MLZ_data}}
\usage{
\S4method{summary}{MLZ_data}(object)
}
\arguments{
\item{object}{An object of class \code{MLZ_data}.}
}
\description{
\code{summary} method for S4 class \code{MLZ_data}
}
\examples{
data(MuttonSnapper)
summary(MuttonSnapper)

}
