% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{summary,MLZ_model-method}
\alias{summary,MLZ_model-method}
\title{\code{summary} method for S4 class \code{MLZ_model}}
\usage{
\S4method{summary}{MLZ_model}(object)
}
\arguments{
\item{object}{An object of class \code{MLZ_model}.}
}
\description{
\code{summary} method for S4 class \code{MLZ_model}
}
\examples{
/dontrun{
data(Goosefish)
goose.model <- ML(Goosefish, ncp = 2, grid.search = FALSE)
summary(goose.model)
}

}
