% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{class}
\name{MLZ_data-class}
\alias{MLZ_data-class}
\title{MLZ_data}
\description{
An S4 class for storing data and life history parameters for a single stock.
Method functions \code{summary} and \code{plot} are available for this class (see examples).
}
\section{Slots}{

\describe{
\item{\code{Stock}}{Name of stock.}

\item{\code{Year}}{A vector of years to be considered in the model. Need to be increments of a single year.}

\item{\code{Len_df}}{A data frame containing individual length observations. The first column should be the Year and the second column should be the length.}

\item{\code{Len_bins}}{A vector of midpoints of length bins for \code{Len_matrix}.}

\item{\code{Len_matrix}}{A matrix of size data. Rows index \code{Year} and columns index \code{Len_bins}.}

\item{\code{vbLinf}}{L-infinity from the von Bertalanffy growth function.}

\item{\code{vbK}}{Parameter K from the von Bertalanffy growth function.}

\item{\code{vbt0}}{Parameter t0 from the von Bertalanffy growth function.}

\item{\code{Lc}}{Length of full selectivity.}

\item{\code{M}}{Natural mortality rate. This is also the lower limit for Z in \code{\link{ML}} and \code{\link{MLCR}}.}

\item{\code{lwb}}{Exponent \code{b} from the allometric length-weight function \eqn{W = aL^b}.}

\item{\code{MeanLength}}{Vector of mean lengths of animals larger than Lc.}

\item{\code{ss}}{Annual sample sizes for MeanLength.}

\item{\code{CPUE}}{Vector of catch-per-unit-effort data.}

\item{\code{Effort}}{Vector of effort data.}

\item{\code{length.units}}{Unit of measurement for lengths, i.e. "cm" or "mm".}
}}

\examples{
data(Goosefish); Goosefish
summary(Goosefish)
plot(Goosefish)

new("MLZ_data")
}
