% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{modal_length}
\alias{modal_length}
\title{Calculate modal length from length data.}
\usage{
modal_length(MLZ_data)
}
\arguments{
\item{MLZ_data}{An object of class \code{MLZ_data}.}
}
\description{
Creates two figures. First, a histogram of length data pooled across all years is produced and the modal
length is reported. Second, a plot of the annual modal length is produced and compared to the pooled modal length.
}
\examples{
data(Nephrops)
modal_length(Nephrops)
}
