% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_MLmulti}
\alias{profile_MLmulti}
\title{Grid search for the multispecies mean length estimator}
\usage{
profile_MLmulti(MLZ.list, ncp, model = c("SSM", "MSM1", "MSM2", "MSM3"),
  spawn = c("continuous", "annual"), stZ = 0.5, parallel = ifelse(ncp > 2,
  TRUE, FALSE), min.time = 3, figure = TRUE, color = TRUE)
}
\arguments{
\item{MLZ.list}{A list containing an object of class \code{MLZ_data} for each species or stock.}

\item{ncp}{The number of change points.}

\item{model}{The name of the multispecies model for the grid search.}

\item{spawn}{Whether 'continuous' or 'annual' (pulse) spawning is modeled.}

\item{stZ}{The starting value of total mortality rate used in the grid search.}

\item{parallel}{Whether the grid search is performed with parallel processing.}

\item{min.time}{The minimum number of years between change points. Only used if \code{ncp > 1}.}

\item{figure}{If \code{TRUE}, creates a plot of the likelihood over the grid search. Only used
if \code{ncp = 1} or \code{2}.}

\item{color}{If \code{TRUE}, creates a color plot for the likelihood surface. Only used if
\code{ncp = 2}.}
}
\description{
A grid search is performed over the time series. Can be used to identify local and global minima.
}
\examples{
data(PRSnapper)
profile_MLmulti(PRSnapper, ncp = 1, model = "MSM1")

}
